refactor: categorized tasks lists code moved to a tasks UI component. feat: component of categorized tasks improved, loading state added
