+++ src/app/app/(home)/page.tsx
@@ -1,44 +1,7 @@
-'use client'
-import { useState, useEffect } from 'react';
-import { TaskFromApi } from '@/app/lib/types/task';
-import { fetchTasksByUserId } from '@/app/lib/utils/api/fetch';
-import { categorizeTasksByTags, mapTaskApiToFrontendTaskType } from '@/app/lib/utils/TasksUtils';
 import { DateNames } from "@/app/lib/utils/DateNames"
-import TasksList from "@/app/ui/tasks/TasksList";
-import tasks from "@/app/lib/mocks/tasks";
-import Tag from "@/app/ui/tasks/Tag";
+import CategorizedTasksLists from "@/app/ui/tasks/CategorizedTasksLists";
 
 export default function Page() {
-  const [tasks, setTasks] = useState<TaskFromApi[]>([]);
-
-  useEffect(() => {
-    const fetchTasks = async (userId: string) => {
-      try {
-        setTasks(await fetchTasksByUserId(userId));
-      } catch (error) {
-        console.error('Error fetching tasks:', error);
-      }
-    };
-
-    const mockUserId = 'cc8fcf59-4708-4ffa-b944-ede6c7816e51';
-    fetchTasks(mockUserId);
-  }, []);
-  
-  const categorizedTasks = categorizeTasksByTags(tasks);
-  const categorizedTasksLists = () => {
-    const tagsTasks = Object.entries(categorizedTasks);
-    const listOfTasksByTag = tagsTasks.map(([tag, tasks]) => {
-      const frontEndTasks = mapTaskApiToFrontendTaskType(tasks);
-      return (
-        <section key={tag}>
-          <Tag name={tag} />
-          <TasksList tasks={frontEndTasks}/>
-        </section>
-      )
-    })
-    return listOfTasksByTag;
-  };
-
   const today = new DateNames(new Date());
 
   return (
@@ -55,7 +18,7 @@ export default function Page() {
         </section>
       </header>
       <section>
-        { categorizedTasksLists() }
+        <CategorizedTasksLists />
       </section>
     </section>
   )
